Fix Bug That Clears Read Status Of Notifications in Notification Controller

+++ Finalproject/Controllers/NotificationsController.cs
@@ -67,12 +67,20 @@ namespace Finalproject.Controllers
         public static async void Update(ApplicationUser currentUser, ApplicationDbContext db)
         {
             var allUserNotifications = db.Notifications.Where(n => n.UserCreator == currentUser).ToList();
+            List<Notification> readNotifications = new List<Notification>();
 
             if (allUserNotifications.Any())
             {
                 foreach (var notification in allUserNotifications)
                 {
-                    db.Notifications.Remove(notification);
+                    if (notification.IsRead == true)
+                    {
+                        readNotifications.Add(notification);
+                    }
+                    else
+                    {
+                        db.Notifications.Remove(notification);
+                    }
                 }
 
                 db.SaveChanges();
@@ -104,8 +112,20 @@ namespace Finalproject.Controllers
                         pyNotification.ProjectId = project.Id;
                         pyNotification.UserCreator = currentUser;
 
-                        db.Notifications.Add(pyNotification);
-                        db.SaveChanges();
+                        bool duplicate = false;
+                        foreach (var notification in readNotifications)
+                        {
+                            if (CheckIfDuplicate(notification, pyNotification, "ProjectNotification") == true)
+                            {
+                                duplicate = true;
+                            }
+                        }
+
+                        if (duplicate == false)
+                        {
+                            db.Notifications.Add(pyNotification);
+                            db.SaveChanges();
+                        }
                     }
                 }
             }
@@ -127,8 +147,20 @@ namespace Finalproject.Controllers
                         taNotification.TaskId = task.Id;
                         taNotification.UserCreator = currentUser;
 
-                        db.Notifications.Add(taNotification);
-                        db.SaveChanges();
+                        bool duplicate = false;
+                        foreach (var notification in readNotifications)
+                        {
+                            if (CheckIfDuplicate(notification, taNotification, "TaskNotification") == true)
+                            {
+                                duplicate = true;
+                            }
+                        }
+
+                        if (duplicate == false)
+                        {
+                            db.Notifications.Add(taNotification);
+                            db.SaveChanges();
+                        }
                     }
                 }
             }
@@ -153,9 +185,35 @@ namespace Finalproject.Controllers
                     taNotification.TaskId = comment.TaskId;
                     taNotification.UserCreator= currentUser;
 
-                    db.Notifications.Add(taNotification);
-                    db.SaveChanges();
+                    bool duplicate = false;
+                    foreach (var notification in readNotifications)
+                    {
+                        if (CheckIfDuplicate(notification, taNotification, "TaskNotification") == true)
+                        {
+                            duplicate = true;
+                        }
+                    }
+
+                    if (duplicate == false)
+                    {
+                        db.Notifications.Add(taNotification);
+                        db.SaveChanges();
+                    }
                 }
+            }
+        }
+
+        public static bool CheckIfDuplicate(Notification notification, Notification newNotification, string discriminator)
+        {
+            if (notification.Title == newNotification.Title &&
+                notification.Description == newNotification.Description &&
+                notification.Discriminator == discriminator)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
             }
         }
     }
